refactor(fuzzing): the sequence generator's four concerns get their own files
The class keeps its lifecycle — boot a throwaway copy, attempt N steps —
and its three tuning knobs. What the domain offers moves to
sequence_generator/catalog.rb, which step to try next to picker.rb, how
a step is built and dispatched to step_builder.rb, and what a success
taught the run to outcome_tracker.rb. All four share the generator's
instance state, so they are included modules rather than objects.
